+++ packages/config/config-core-sync/lib/index.ts
@@ -1,5 +1,0 @@
-export * from './Config';
-export * from './ConfigSource';
-export * from './Api';
-export * from './EnvSource';
-export * from './reactive';