+++ packages/config/config-core/lib/ReactiveEmitter/index.ts
@@ -1,0 +1,2 @@
+export * from "./ReactiveEmitter";
+export * from "./reactive";