+++ packages/config/config-core/lib/source/index.ts
@@ -1,0 +1,2 @@
+export * from "./SourceManager";
+export * from "./source";