+++ packages/config/config-core/src/index.ts
@@ -1,27 +1,8 @@
-import { Config, EnvSource, ConfigEvents } from '../lib';
+import { Config } from '../lib';
 
-class TheEnvSource extends EnvSource {
-  initEnv(): boolean {
-    return true;
-  }
-  getEnv(): Record<string, any> {
-    return {
-      a__bbb: '666',
-      a__bb: '888',
-      a__cc: '777',
-    };
-  }
-  reset() {
-    console.log(`123`, 123);
-  }
-  save(path: string, value: any) {
-    console.log(`save path=${path} value=${value}`);
-  }
-}
 async function bootstrap() {
   const m = await Config.create({
     sources: [
-      new TheEnvSource(),
       {
         load() {
           return {
@@ -33,13 +14,17 @@ async function bootstrap() {
             },
           };
         },
+        save() {
+          console.log(`12231`, 12231)
+        }
       },
     ],
   });
   // m.on(ConfigEvents.change, (config) => {
   //   console.log(`config change`, config);
   // });
-  m.on(ConfigEvents.valueChange, ({ path, value }) => console.log(`onvaluechange path,value`, path, value));
+  m.on('change', ({ path, value }) => console.log(`onvaluechange path,value`, path, value));
+  m.set('a', 'bbb')
   // console.log(`获取全部 m.get()`, m.get());
   // // console.log(`m.get()`, JSON.stringify(m.get()));
   // console.log(`通过路径获取 m.get('a.b.c')`, m.get('a.b.c'));
@@ -54,9 +39,6 @@ async function bootstrap() {
   // console.log(`修改nnn为 1234后 m.get('nnn')`, m.get('nnn'));
   // await m.reset();
   // console.log(`m.config`, m.config);
-  m.config.arr.push({a:1});
-  m.config.arr.push('3');
-  m.config.arr[0].a = 2;
-  console.log(`m.config.arr`, m.config.arr);
+
 }
 bootstrap();