+++ packages/config/config-core/tsup.config.ts
@@ -3,7 +3,7 @@ import packageJson from './package.json';
 const { name, version } = packageJson;
 console.log(JSON.stringify({ name, version }));
 export default defineConfig({
-  entry: ['lib/index.ts'],
+  entry: ['lib/index.ts', '!**/*.test.ts'],
   format: ['cjs', 'esm'],
   dts: true,
   splitting: true,