+++ packages/config/config-for-web/lib/ViteEnvSource.ts
@@ -1,4 +1,4 @@
-import { EnvSource } from "@tikkhun/config-core";
+import { type EnvSource } from "@tikkhun/config-core";
 export class ViteEnvSource extends EnvSource {
   getEnv() {
     return import.meta.env;