+++ packages/utils/node/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tikkhun/utils
 
+## 1.0.18
+
+### Patch Changes
+
+- e50d6d5: feat: findUp file 以防在子目录上运行没有效果
+
 ## 1.0.17
 
 ### Patch Changes