+++ tools/release-plugins/info-store/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tikkhun/release-info-store
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [e50d6d5]
+  - @tikkhun/info@0.0.6
+
 ## 0.0.3
 
 ### Patch Changes